Make the existing devchat site more general
Use the first of element that's either audio or video. This makes this
site usable with a 'naked' stream site (just an <audio> element pointed
at a stream), like this one:

http://brokenbeats.net:8000/tune

With this change, you can just set aliases on devchat.tv for the sites
you want ("brokenbeats" will work for this example).

A future change would be to rename the files and the site list entry to
indicate how generally applicable this site is.