Don't create a controller if there's no player
Previously, errors reading "Cannot read property 'play' of undefined"
were logged when no player was present. Instead, don't create a
controller at all: Check for a player first.

This also makes it so that pages without a player won't be listed in the
StreamKeys UI, which makes sense.

(This was a defect in the original devchat controller, and not
introduced by the refactor.)