Generic single embedded element controller (#466)
* Reinstate devchat

This reverts commit c61af38c2727a5a95624a3f10a5458700dd93fe2.

It looks like the devchat controller was removed to avoid jQuery in the
code; reverting that change first, to form a base for additional changes
to reinstante the controller without jQuery.

* Refactor devchat controller

Pull the inject code into the main controller, and do away with the
custom inject code entirely.

Remove all references to jQuery.

Remove the custom 'sk-not' class and use native element controls
instead.

* Use the page's title as the song name

* Don't create a controller if there's no player

Previously, errors reading "Cannot read property 'play' of undefined"
were logged when no player was present. Instead, don't create a
controller at all: Check for a player first.

This also makes it so that pages without a player won't be listed in the
StreamKeys UI, which makes sense.

(This was a defect in the original devchat controller, and not
introduced by the refactor.)

* Rename to Embedded Element

This controller doesn't actually _work_ with devchat.tv anymore, due to
changes in their site. It does, however, work for any site with an
embedded <audio> or <video> element. Rebrand appropriately.

* Add an alias

This passes unit tests.

In real life, example.com won't ever be a site used, but since I set it
as the url, I have to put it in here. I'm open to suggestions as to how
to correctly handle this: Maybe set a canonical example site instead?

* Remove url from embedded site definiton

Scope unit tests to only test sites with urls; a null url shouldn't be
tested, since there isn't a url to test it against.