youtube: Fix detection of playing status
The selector for YouTube playState (play button) used to be
".ytp-play-button[aria-label='Pause']" but I recently found that it was
changed to ".ytp-play-button[aria-label='Pause (k)']" in the source of
the YouTube page, breaking the detection of the YouTube playing state in
StreamKeys. Fix that by checking if 'aria-label' begins with 'Pause'
instead of using an exact match.

Signed-off-by: Alex Tsitsimpis <alex.tsitsimpis@gmail.com>